+++ packages/react/src/components/date-picker/date-picker.decorator.tsx
@@ -1,0 +1,8 @@
+import { DatePickerParameters } from '@dso-toolkit/sources';
+import * as React from 'react';
+
+export const decorator: DatePickerParameters<JSX.Element>['decorator'] = (story) => (
+  <div style={({ width: '175px' })}>
+    {story()}
+  </div>
+);