+++ packages/sources/src/components/date-picker/date-picker.models.ts
@@ -1,6 +1,7 @@
 export interface DatePicker {
   id?: string;
   onDateChange: (e: CustomEvent) => void;
+  direction?: string;
   value?: string;
   min?: number;
   max?: number;